refactors Mvc directory and Middleware
