+++ src/Handlers/VIewCacheHandler.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Quantum\Handlers;
+
+use Quantum\Http\Response;
+use Quantum\Libraries\ResourceCache\ViewCache;
+
+class ViewCacheHandler
+{
+    protected ViewCache $viewCache;
+
+    public function __construct()
+    {
+        $this->viewCache = ViewCache::getInstance();
+    }
+
+    public function serveCachedView(string $uri, Response $response): bool
+    {
+        if ($this->viewCache->isEnabled() && $this->viewCache->exists($uri)) {
+            $response->html($this->viewCache->get($uri));
+            return true;
+        }
+
+        return false;
+    }
+}
